% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{get_reference}
\alias{get_reference}
\title{Get reference data frame}
\usage{
get_reference(context, epitope = NULL, org = NULL, publication = NULL)
}
\arguments{
\item{context}{The context to get the reference data frame for}

\item{epitope}{The epitope to get the reference data frame for}

\item{org}{The organism to get the reference data frame for}

\item{publication}{The publication to get the reference data frame for}
}
\value{
A data frame with the reference data
}
\description{
Get reference data frame
}
