% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_cols.R
\name{lookup_cols}
\alias{lookup_cols}
\title{Lookup columns to match}
\usage{
lookup_cols(
  query,
  reference,
  CDRH3,
  VH = NA,
  JH = NA,
  CDRL3 = NA,
  VL = NA,
  JL = NA
)
}
\arguments{
\item{query}{The query data frame}

\item{reference}{The reference data frame}

\item{CDRH3}{The heavy chain CDR3 sequence}

\item{VH}{The heavy chain V gene}

\item{JH}{The heavy chain J gene}

\item{CDRL3}{The light chain CDR3 sequence}

\item{VL}{The light chain V gene}

\item{JL}{The light chain J gene}
}
\value{
A list of columns to match
}
\description{
Lookup columns to match
}
